chore(parser): Split parser file in multiple files
Parser file was 1651 lines long so it was time to split

Related to github issue #25
