+++ parser/parser.go
@@ -225,221 +225,6 @@ func (p *Parser) ParseFile() *ast.File {
 	return f
 }
 
-// parseFuncDecl returns function declaration
-func (p *Parser) parseFuncDecl() ast.Decl {
-	kw := p.expect(token.KWFunc, "expected 'func'")
-	name := p.expect(token.Ident, "expected function name")
-	_ = p.expect(token.LParen, "expected '(' after function name")
-
-	f := &ast.FuncDecl{
-		FuncKW: kw,
-		Name:   name,
-	}
-	for p.kind() != token.RParen && p.kind() != token.EOF {
-		if p.kind() == token.Comma {
-			tok := p.next()
-			p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", tok.Line, tok.Column, tok.Kind, tok.Value))
-			return &ast.BadDecl{From: kw, To: tok, Reason: "expected expression not ','"}
-		}
-		f.Params = append(f.Params, p.parseFuncParam())
-		if p.kind() != token.Comma && p.kind() != token.RParen && p.kind() != token.EOF {
-			tok := p.next()
-			p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", tok.Line, tok.Column, tok.Kind, tok.Value))
-			return &ast.BadDecl{From: kw, To: tok, Reason: "expected ',' or ')'"}
-		}
-
-		if p.kind() == token.Comma {
-			_ = p.next()
-			if p.kind() == token.RParen || p.kind() == token.EOF {
-				p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-				return &ast.BadDecl{From: kw, To: p.peek(), Reason: "expected expression after ','"}
-			}
-		}
-	}
-
-	_ = p.expect(token.RParen, "expected ')' after function name")
-
-	f.Results = p.parseFuncReturnTypes()
-	body := p.parseBlock()
-	f.Body = body
-
-	return f
-}
-
-// parseFuncParam returns function parameter
-func (p *Parser) parseFuncParam() ast.Param {
-	name := p.expect(token.Ident, "expected parameter identifier")
-	typ, btyp, bad := p.parseFuncParamType()
-	if bad {
-		return ast.Param{Name: name, Type: btyp}
-	}
-	return ast.Param{Name: name, Type: typ}
-}
-
-// parseFuncParamType returns func parameter type
-func (p *Parser) parseFuncParamType() (*ast.NameType, *ast.BadType, bool) {
-	typ := &ast.NameType{}
-	btyp := &ast.BadType{}
-	var bad bool
-	tok := p.peek()
-
-	if token.IsFuncParamTypes(tok.Kind) {
-		typ.Name = tok
-	} else {
-		p.errors = append(p.errors, fmt.Errorf("%d:%d: unsupported type with %v %q", tok.Line, tok.Column, tok.Kind, tok.Value))
-		btyp.From = tok
-		btyp.Reason = "unexpected type name"
-		bad = true
-	}
-
-	_ = p.next()
-	return typ, btyp, bad
-}
-
-// parseFuncReturnTypes returns func return types
-func (p *Parser) parseFuncReturnTypes() ast.ReturnTypes {
-	if p.kind() == token.LBrace {
-		return ast.ReturnTypes{}
-	}
-
-	var result ast.ReturnTypes
-	if p.kind() == token.LParen {
-		lp := p.expect(token.LParen, "expected '('")
-		if p.kind() == token.RParen {
-			p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-			btyp := &ast.BadType{
-				From:   lp,
-				To:     p.peek(),
-				Reason: "expected parameter(s) before ')'",
-			}
-			result.List = append(result.List, ast.Param{Name: p.peek(), Type: btyp})
-			return result
-		}
-
-		if p.kind() == token.Comma {
-			p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-			btyp := &ast.BadType{
-				From:   lp,
-				To:     p.peek(),
-				Reason: "expected expression before ','",
-			}
-
-			result.List = append(result.List, ast.Param{Name: p.peek(), Type: btyp})
-			return result
-		}
-
-		result.LParen = lp
-		// entering into kind: (indentA indentB, indentC indentD) or (indentA indentB)
-		if p.kindNext(p.position+2) == token.Comma || p.kindNext(p.position+2) == token.RParen {
-			for p.kind() != token.RParen && p.kind() != token.LBrace && p.kind() != token.EOF {
-				param := p.parseFuncParam()
-				_, bad := param.Type.(*ast.BadType)
-				if bad {
-					result.List = append(result.List, param)
-					return result
-				}
-				result.List = append(result.List, param)
-
-				if p.kind() != token.Comma && p.kind() != token.RParen {
-					p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-					btyp := &ast.BadType{
-						From:   lp,
-						To:     p.peek(),
-						Reason: "expected ',' after parameter(s)",
-					}
-					result.List = append(result.List, ast.Param{Name: p.peek(), Type: btyp})
-					return result
-				}
-
-				if p.kind() == token.Comma {
-					comma := p.expect(token.Comma, "expected ','")
-					if p.kind() == token.RParen {
-						p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-						btyp := &ast.BadType{
-							From:   comma,
-							To:     p.peek(),
-							Reason: "expected parameter(s) after ','",
-						}
-						result.List = append(result.List, ast.Param{Name: p.peek(), Type: btyp})
-						return result
-					}
-				}
-			}
-		} else {
-			// entering into kind: (type, type)
-			for p.kind() != token.RParen && p.kind() != token.LBrace && p.kind() != token.EOF {
-				typ, btyp, bad := p.parseFuncParamType()
-				if bad {
-					result.List = append(result.List, ast.Param{Type: btyp})
-					return result
-				}
-				result.List = append(result.List, ast.Param{Type: typ})
-
-				if p.kind() != token.Comma && p.kind() != token.RParen {
-					p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-					btyp := &ast.BadType{
-						From:   lp,
-						To:     p.peek(),
-						Reason: "expected ',' after parameter(s)",
-					}
-					result.List = append(result.List, ast.Param{Name: p.peek(), Type: btyp})
-					return result
-				}
-
-				if p.kind() == token.Comma {
-					comma := p.expect(token.Comma, "expected ','")
-					if p.kind() == token.RParen {
-						p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-						btyp := &ast.BadType{
-							From:   comma,
-							To:     p.peek(),
-							Reason: "expected parameter(s) after ','",
-						}
-						result.List = append(result.List, ast.Param{Name: p.peek(), Type: btyp})
-						return result
-					}
-				}
-			}
-		}
-
-		rp := p.expect(token.RParen, "expected ')'")
-		result.RParen = rp
-		return result
-	}
-
-	if p.kind() == token.Comma {
-		p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-		btyp := &ast.BadType{
-			From:   p.peek(),
-			Reason: "expected builtin type or ident",
-		}
-		result.List = append(result.List, ast.Param{Name: p.peek(), Type: btyp})
-		return result
-	}
-
-	next := p.peek()
-	typ, btyp, bad := p.parseFuncParamType()
-	if bad {
-		p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", btyp.From.Line, btyp.From.Column, btyp.From.Kind, btyp.From.Value))
-		result.List = append(result.List, ast.Param{Type: btyp})
-		return result
-	}
-	result.List = append(result.List, ast.Param{Type: typ})
-
-	if p.kind() != token.LBrace {
-		p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-		btyp := &ast.BadType{
-			From:   next,
-			To:     p.peek(),
-			Reason: "expected '{' after type",
-		}
-		result.List = append(result.List, ast.Param{Name: p.peek(), Type: btyp})
-		return result
-	}
-
-	return result
-}
-
 // parseBlock returns declaration within curly braces
 func (p *Parser) parseBlock() *ast.BlockStmt {
 	lb := p.expect(token.LBrace, "expected '{'")
@@ -561,69 +346,6 @@ func (p *Parser) parseSimpleStmt() ast.Stmt {
 		return &ast.BadStmt{From: left.Start(), Reason: "unsupported statement"}
 	}
 	return &ast.ExprStmt{Expr: left}
-}
-
-// parseConstDecl returns constant declaration
-func (p *Parser) parseConstDecl() ast.Stmt {
-	kw := p.expect(token.KWConst, "expected 'const'")
-	name := p.expect(token.Ident, "expected constant name")
-
-	typ, btyp, bad := p.parseVarConstType()
-	if bad {
-		p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", btyp.From.Line, btyp.From.Column, btyp.From.Kind, btyp.From.Value))
-		return btyp
-	}
-	eq := p.expect(token.Assign, "expected '=")
-	init := p.parseExpr(LOWEST)
-
-	return &ast.ConstDeclStmt{
-		ConstKW: kw,
-		Name:    name,
-		Type:    typ,
-		Eq:      eq,
-		Init:    init,
-	}
-}
-
-// parseVarDecl returns variable declaration
-func (p *Parser) parseVarDecl() ast.Stmt {
-	kw := p.expect(token.KWVar, "expected 'var'")
-	name := p.expect(token.Ident, "expected variable name")
-
-	typ, btyp, bad := p.parseVarConstType()
-	if bad {
-		return btyp
-	}
-	eq := p.expect(token.Assign, "expected '=")
-	init := p.parseExpr(LOWEST)
-
-	return &ast.VarDeclStmt{
-		VarKW: kw,
-		Name:  name,
-		Type:  typ,
-		Eq:    eq,
-		Init:  init,
-	}
-}
-
-// parseVarConstType returns const/vars types
-func (p *Parser) parseVarConstType() (*ast.NameType, *ast.BadType, bool) {
-	typ := &ast.NameType{}
-	btyp := &ast.BadType{}
-	var bad bool
-	tok := p.peek()
-
-	if token.IsVarConstTypes(tok.Kind) {
-		typ.Name = tok
-	} else {
-		p.errors = append(p.errors, fmt.Errorf("%d:%d: unsupported type with %v %q", tok.Line, tok.Column, tok.Kind, tok.Value))
-		btyp.From = tok
-		btyp.Reason = "unexpected type name"
-		bad = true
-	}
-
-	_ = p.next()
-	return typ, btyp, bad
 }
 
 // parseExpr returns the type of declaration being parsed
@@ -822,830 +544,8 @@ func (p *Parser) parseStmtExpr(left ast.Expr) *ast.AssignStmt {
 	}
 }
 
-// parseIncDecStmtExpr returns expressions for parseStmt func
-func (p *Parser) parseIncDecStmtExpr(left ast.Expr) *ast.IncDecStmt {
-	op := p.peek()
-	_ = p.next()
-	return &ast.IncDecStmt{
-		X:        left,
-		Operator: op,
-	}
-}
-
-// parseReturnStmtExpr returns expressions for parseStmt func
-func (p *Parser) parseReturnStmtExpr() ast.Stmt {
-	rn := p.expect(token.KWReturn, "expected 'return'")
-
-	// return has no values
-	if p.kind() == token.EOF || p.kind() == token.RBrace {
-		return &ast.ReturnStmt{
-			Return: rn,
-		}
-	}
-
-	var args []ast.Expr
-	for p.kind() != token.RBrace && p.kind() != token.EOF {
-		if p.kind() == token.Comma {
-			tok := p.next()
-			p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", tok.Line, tok.Column, tok.Kind, tok.Value))
-			return &ast.BadStmt{From: rn, To: tok, Reason: "expected expression not ','"}
-		}
-
-		args = append(args, p.parseExpr(LOWEST))
-		if p.kind() == token.KWCase || p.kind() == token.KWDefault {
-			break
-		}
-
-		if p.kind() != token.Comma && p.kind() != token.RBrace && p.kind() != token.EOF {
-			tok := p.next()
-			p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", tok.Line, tok.Column, tok.Kind, tok.Value))
-			return &ast.BadStmt{From: rn, To: tok, Reason: "expected ',' or '}' after return value"}
-		}
-
-		if p.kind() == token.Comma {
-			_ = p.next()
-			if p.kind() == token.RBrace || p.kind() == token.EOF {
-				p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-				return &ast.BadStmt{From: rn, To: p.peek(), Reason: "expected expression after ','"}
-			}
-		}
-	}
-
-	return &ast.ReturnStmt{
-		Return: rn,
-		Values: args,
-	}
-}
-
-// parseIfStmtExpr returns expressions for parseStmt func
-func (p *Parser) parseIfStmtExpr() ast.Stmt {
-	ifs := p.expect(token.KWIf, "expected 'if'")
-	stmt := &ast.IfStmt{
-		If: ifs,
-	}
-
-	if p.kind() == token.LBrace {
-		p.errors = append(p.errors, fmt.Errorf("%d:%d: missing condition, got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-		return &ast.BadStmt{From: ifs, To: p.peek(), Reason: "missing condition after 'if'"}
-	}
-
-	stmt.Condition = p.parseExpr(LOWEST)
-	if token.IsAssignment(p.kind()) {
-		tok := p.next()
-		p.errors = append(p.errors, fmt.Errorf("%d:%d: assignment not allowed in if condition, got %v %q", tok.Line, tok.Column, tok.Kind, tok.Value))
-		return &ast.BadStmt{From: ifs, To: tok, Reason: "assignment not allowed in if condition; use =="}
-	}
-
-	stmt.Then = p.parseBlock()
-
-	if p.kind() == token.KWElse {
-		_ = p.next()
-		if p.kind() == token.KWIf {
-			stmt.Else = p.parseIfStmtExpr()
-		} else if p.kind() == token.LBrace {
-			stmt.Else = p.parseBlock()
-		} else {
-			tok := p.next()
-			p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", tok.Line, tok.Column, tok.Kind, tok.Value))
-			return &ast.BadStmt{From: ifs, To: tok, Reason: "expected expression '{' or 'if' after 'else'"}
-		}
-	}
-
-	return stmt
-}
-
-// parseForStmtExpr returns expressions for parseStmt func
-func (p *Parser) parseForStmtExpr() ast.Stmt {
-	ftok := p.expect(token.KWFor, "expected 'for'")
-	fstmt := &ast.ForStmt{
-		For: ftok,
-	}
-	rstmt := &ast.RangeStmt{
-		For: ftok,
-	}
-
-	// infinite loop
-	if p.kind() == token.LBrace {
-		fstmt.Body = p.parseForBlockStmt()
-		return fstmt
-	}
-
-	if p.lookForInForHeader(token.KWRange) {
-		// for range x {}
-		if p.kind() == token.KWRange {
-			rg := p.expect(token.KWRange, "expected 'range'")
-			rstmt.Range = rg
-			if p.kind() == token.LBrace {
-				p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-				return &ast.BadStmt{From: ftok, To: p.peek(), Reason: "expected expression before '{'"}
-			}
-			rstmt.X = p.parseExpr(LOWEST)
-
-			if p.kind() != token.LBrace {
-				p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-				return &ast.BadStmt{From: ftok, To: p.peek(), Reason: "expected '{' after expression"}
-			}
-
-			rstmt.Body = p.parseForBlockStmt()
-			return rstmt
-		}
-
-		// var k1, k2 ast.Expr
-		if p.kind() == token.Comma {
-			tok := p.next()
-			p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", tok.Line, tok.Column, tok.Kind, tok.Value))
-			return &ast.BadStmt{From: ftok, To: tok, Reason: "expected expression not ','"}
-		}
-
-		k1 := p.expect(token.Ident, "expected 'identifier'")
-		if k1.Kind != token.Ident {
-			return &ast.BadStmt{From: ftok, To: k1, Reason: "expected identifier"}
-		}
-		xk1 := &ast.IdentExpr{Name: k1}
-
-		// for k,v := range x {}
-		if p.kind() == token.Comma {
-			_ = p.next()
-			rstmt.Key = xk1
-			k2 := p.expect(token.Ident, "expected 'identifier'")
-			if k2.Kind != token.Ident {
-				return &ast.BadStmt{From: ftok, To: k2, Reason: "expected identifier"}
-			}
-
-			rstmt.Value = &ast.IdentExpr{Name: k2}
-			if !token.IsRangeForAssignment(p.kind()) {
-				p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-				return &ast.BadStmt{From: ftok, To: p.peek(), Reason: "expected '=' or ':=' after expression"}
-			}
-			op := p.next()
-			rstmt.Op = op
-			rg := p.expect(token.KWRange, "expected 'range'")
-			rstmt.Range = rg
-			if p.kind() == token.LBrace {
-				p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-				return &ast.BadStmt{From: ftok, To: p.peek(), Reason: "expected expression before '{'"}
-			}
-			rstmt.X = p.parseExpr(LOWEST)
-
-			if p.kind() != token.LBrace {
-				p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-				return &ast.BadStmt{From: ftok, To: p.peek(), Reason: "expected '{' after expression"}
-			}
-
-			rstmt.Body = p.parseForBlockStmt()
-			return rstmt
-		}
-
-		// for v := range x {}
-		if !token.IsRangeForAssignment(p.kind()) {
-			p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-			return &ast.BadStmt{From: ftok, To: p.peek(), Reason: "expected '=' or ':=' after expression"}
-		}
-
-		rstmt.Key = xk1
-		op := p.next()
-		rstmt.Op = op
-		rg := p.expect(token.KWRange, "expected 'range'")
-		rstmt.Range = rg
-		if p.kind() == token.LBrace {
-			p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-			return &ast.BadStmt{From: ftok, To: p.peek(), Reason: "expected expression before '{'"}
-		}
-		rstmt.X = p.parseExpr(LOWEST)
-
-		if p.kind() != token.LBrace {
-			p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-			return &ast.BadStmt{From: ftok, To: p.peek(), Reason: "expected '{' after expression"}
-		}
-
-		rstmt.Body = p.parseForBlockStmt()
-		return rstmt
-	}
-
-	// for condition {}
-	if !p.lookForInForHeader(token.SemiComma) {
-		fstmt.Condition = p.parseExpr(LOWEST)
-		if p.kind() != token.LBrace {
-			p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-			return &ast.BadStmt{From: ftok, To: p.peek(), Reason: "expected '{' after condition"}
-		}
-
-		fstmt.Body = p.parseForBlockStmt()
-		return fstmt
-	}
-
-	// for init; condition; post {}
-	fstmt.Init = p.parseSimpleStmt()
-	if !isValidForInit(fstmt.Init) {
-		p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected init statement in for header, got %v %q", fstmt.Init.Start().Line, fstmt.Init.Start().Column, fstmt.Init.Start().Kind, fstmt.Init.Start().Value))
-		return &ast.BadStmt{From: fstmt.Init.Start(), To: fstmt.Init.End(), Reason: "invalid init statement in for header"}
-	}
-	sm1 := p.expect(token.SemiComma, "expected ';'")
-	if sm1.Kind != token.SemiComma {
-		return &ast.BadStmt{From: ftok, To: sm1, Reason: "expected ';'"}
-	}
-
-	fstmt.Condition = p.parseExpr(LOWEST)
-	sm2 := p.expect(token.SemiComma, "expected ';'")
-	if sm2.Kind != token.SemiComma {
-		return &ast.BadStmt{From: ftok, To: sm2, Reason: "expected ';'"}
-	}
-
-	if p.kind() == token.LBrace {
-		p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-		return &ast.BadStmt{From: ftok, To: p.peek(), Reason: "expected expression before '{'"}
-	}
-
-	fstmt.Post = p.parseSimpleStmt()
-	if !isValidForPost(fstmt.Post) {
-		p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected post statement in for header, got %v %q", fstmt.Post.Start().Line, fstmt.Post.Start().Column, fstmt.Post.Start().Kind, fstmt.Post.Start().Value))
-		return &ast.BadStmt{From: fstmt.Post.Start(), To: fstmt.Post.End(), Reason: "invalid post statement in for header"}
-	}
-	if p.kind() != token.LBrace {
-		tok := p.next()
-		p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", tok.Line, tok.Column, tok.Kind, tok.Value))
-		return &ast.BadStmt{From: fstmt.Post.Start(), To: tok, Reason: "expected '{'"}
-	}
-
-	fstmt.Body = p.parseForBlockStmt()
-	return fstmt
-}
-
-// isValidForInit is only used in for init statement
-func isValidForInit(s ast.Stmt) bool {
-	switch st := s.(type) {
-	case *ast.AssignStmt:
-		return true
-	case *ast.ExprStmt:
-		_, ok := st.Expr.(*ast.CallExpr)
-		return ok
-	default:
-		return false
-	}
-}
-
-// isValidForPost is only used in for post statement
-func isValidForPost(s ast.Stmt) bool {
-	switch st := s.(type) {
-	case *ast.AssignStmt, *ast.IncDecStmt:
-		return true
-	case *ast.ExprStmt:
-		_, ok := st.Expr.(*ast.CallExpr)
-		return ok
-	default:
-		return false
-	}
-}
-
-// parseForBlockStmt is an helper to make sure we inc/dec
-// loopDepth counter
-func (p *Parser) parseForBlockStmt() *ast.BlockStmt {
-	p.loopDepth++
-	defer func() {
-		p.loopDepth--
-	}()
-	return p.parseBlock()
-}
-
-// parseBreakStmt returns expressions for parseStmt func
-func (p *Parser) parseBreakStmt() ast.Stmt {
-	if p.loopDepth == 0 {
-		tok := p.next()
-		p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected break expression outside for loop, got %v %q", tok.Line, tok.Column, tok.Kind, tok.Value))
-		return &ast.BadStmt{From: tok, Reason: "expected 'break' inside 'for' loop"}
-	}
-
-	kw := p.expect(token.KWBreak, "expected 'break'")
-	if p.kind() == token.Comment {
-		_ = p.next()
-	}
-	// we do not accept labels for now so anything unauthorized is rejected
-	// maybe labels will be supported later
-	if p.kind() == token.RBrace || p.kind() == token.EOF || p.kind() == token.SemiComma || p.peek().Line > kw.Line {
-		if p.kind() == token.SemiComma {
-			_ = p.next()
-		}
-		return &ast.BreakStmt{
-			Break: kw,
-		}
-	}
-
-	p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected statement after 'break', got %v %q", kw.Line, kw.Column, p.peek().Kind, p.peek().Value))
-	return &ast.BadStmt{From: p.peek(), Reason: "expected '}' or 'EOF' or new line"}
-}
-
-// parseContinueStmt returns expressions for parseStmt func
-func (p *Parser) parseContinueStmt() ast.Stmt {
-	if p.loopDepth == 0 {
-		tok := p.next()
-		p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected continue expression outside for loop, got %v %q", tok.Line, tok.Column, tok.Kind, tok.Value))
-		return &ast.BadStmt{From: tok, Reason: "expected 'continue' inside 'for' loop"}
-	}
-
-	kw := p.expect(token.KWContinue, "expected 'continue'")
-	if p.kind() == token.Comment {
-		_ = p.next()
-	}
-	// any unauthorized statement is rejected after 'continue'
-	if p.kind() == token.RBrace || p.kind() == token.EOF || p.kind() == token.SemiComma || p.peek().Line > kw.Line {
-		if p.kind() == token.SemiComma {
-			_ = p.next()
-		}
-		return &ast.ContinueStmt{
-			Continue: kw,
-		}
-	}
-
-	p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected statement after 'continue', got %v %q", kw.Line, kw.Column, p.peek().Kind, p.peek().Value))
-	return &ast.BadStmt{From: p.peek(), Reason: "expected '}' or 'EOF' or new line"}
-}
-
-func (p *Parser) parseSwitchStmt() ast.Stmt {
-	kw := p.expect(token.KWSwitch, "expected 'switch'")
-	s := ast.SwitchStmt{
-		Switch: kw,
-	}
-
-	if p.lookForInSwitchHeader(token.SemiComma) {
-		s.Init = p.parseSimpleStmt()
-		if p.kind() == token.SemiComma {
-			_ = p.next()
-		}
-
-		if p.kind() != token.LBrace {
-			s.Tag = p.parseExpr(LOWEST)
-		}
-		lb := p.expect(token.LBrace, "expected '{'")
-		s.LBrace = lb
-
-		return p.parseSwitchCasesStmt(s)
-	}
-
-	if p.kind() != token.LBrace {
-		s.Tag = p.parseExpr(LOWEST)
-	}
-	lb := p.expect(token.LBrace, "expected '{'")
-	s.LBrace = lb
-
-	return p.parseSwitchCasesStmt(s)
-}
-
-func (p *Parser) parseSwitchCasesStmt(s ast.SwitchStmt) ast.Stmt {
-	var dcount int
-	for p.kind() != token.RBrace && p.kind() != token.EOF {
-		switch p.kind() {
-		case token.KWCase:
-			ckw := p.expect(token.KWCase, "expected 'case'")
-			scase := ast.CaseClause{
-				Case: ckw,
-			}
-
-			if p.lookForInSwitchCaseHeader(token.Comma) {
-				for p.kind() != token.Colon && p.kind() != token.RBrace && p.kind() != token.EOF {
-					if p.kind() == token.Comma {
-						p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-						return &ast.BadStmt{From: s.Switch, To: p.peek(), Reason: "expected expression before ','"}
-					}
-
-					scase.Values = append(scase.Values, p.parseExpr(LOWEST))
-					if p.kind() == token.Colon {
-						colon := p.expect(token.Colon, "expected ':'")
-						scase.Colon = colon
-
-						for p.kind() != token.KWCase && p.kind() != token.KWDefault && p.kind() != token.RBrace && p.kind() != token.EOF {
-							scase.Body = append(scase.Body, p.parseStmt())
-						}
-						break
-					}
-
-					if p.kind() != token.Comma && p.kind() != token.Colon {
-						p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-						return &ast.BadStmt{From: s.Switch, To: p.peek(), Reason: "expected ','"}
-					}
-					_ = p.expect(token.Comma, "expected ','")
-				}
-			} else {
-				for p.kind() != token.Colon && p.kind() != token.RBrace && p.kind() != token.EOF {
-					scase.Values = append(scase.Values, p.parseExpr(LOWEST))
-					if p.kind() == token.Colon {
-						colon := p.expect(token.Colon, "expected ':'")
-						scase.Colon = colon
-
-						for p.kind() != token.KWCase && p.kind() != token.KWDefault && p.kind() != token.RBrace && p.kind() != token.EOF {
-							scase.Body = append(scase.Body, p.parseStmt())
-						}
-						break
-					}
-
-					if p.kind() != token.Colon {
-						p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-						return &ast.BadStmt{From: s.Switch, To: p.peek(), Reason: "expected ','"}
-					}
-				}
-			}
-			s.Cases = append(s.Cases, scase)
-
-		case token.KWDefault:
-			dcount++
-			if dcount > 1 {
-				p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected 'default', got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-				return &ast.BadStmt{From: s.Switch, To: p.peek(), Reason: "expected only one 'default' case"}
-			}
-
-			dkw := p.expect(token.KWDefault, "expected 'default'")
-			if p.kind() == token.Colon {
-				colon := p.expect(token.Colon, "expected ':'")
-				dcase := ast.CaseClause{
-					Case:  dkw,
-					Colon: colon,
-				}
-
-				// empty body
-				if p.kind() == token.KWCase {
-					s.Cases = append(s.Cases, dcase)
-				} else {
-					for p.kind() != token.KWCase && p.kind() != token.KWDefault && p.kind() != token.RBrace && p.kind() != token.EOF {
-						dcase.Body = append(dcase.Body, p.parseStmt())
-					}
-					s.Cases = append(s.Cases, dcase)
-				}
-			}
-
-			if p.kind() != token.KWDefault && p.kind() != token.KWCase && p.kind() != token.RBrace && p.kind() != token.EOF {
-				p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected ':', got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-				return &ast.BadStmt{From: s.Switch, To: p.peek(), Reason: "expected ':' after 'default'"}
-			}
-
-		default:
-			p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-			return &ast.BadStmt{From: s.Switch, To: p.peek(), Reason: "expected 'case' or 'default'"}
-		}
-	}
-
-	rb := p.expect(token.RBrace, "expected '}'")
-	s.RBrace = rb
-
-	return &s
-}
-
-// parseFallThroughStmt returns expressions for parseStmt func
-func (p *Parser) parseFallThroughStmt() ast.Stmt {
-	kw := p.expect(token.KWFallThrough, "expected 'fallthrough'")
-	if p.kind() == token.Comment {
-		_ = p.next()
-	}
-
-	// any unauthorized statement is rejected after 'fallthrough'
-	if p.kind() == token.RBrace || p.kind() == token.EOF || p.kind() == token.KWCase || p.kind() == token.KWDefault {
-		return &ast.FallThroughStmt{
-			FallThroughStmt: kw,
-		}
-	}
-
-	p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected statement after 'fallthrough', got %v %q", kw.Line, kw.Column, p.peek().Kind, p.peek().Value))
-	return &ast.BadStmt{From: p.peek(), Reason: "expected '}' or 'EOF' or new line"}
-}
-
-// parseStructType returns parsed struct
-func (p *Parser) parseStructType() *ast.StructType {
-	kwt := p.expect(token.KWType, "expected 'type'")
-	kwi := p.expect(token.Ident, "expected 'ident'")
-	kws := p.expect(token.KWStruct, "expected 'struct'")
-	lbrace := p.expect(token.LBrace, "expected '{'")
-
-	st := &ast.StructType{
-		TypeDecl: kwt,
-		Name:     kwi,
-		Public:   isPublic(kwi),
-		Struct:   kws,
-		LBrace:   lbrace,
-	}
-
-	for p.kind() != token.RBrace && p.kind() != token.EOF {
-		st.Fields = append(st.Fields, p.parseStructTypeField())
-
-		if p.kind() == token.Comment {
-			_ = p.next()
-		}
-
-		if p.kind() == token.SemiComma {
-			_ = p.next()
-			continue
-		}
-
-		if p.kind() == token.RBrace {
-			break
-		}
-
-		if p.newlineSincePrev() {
-			continue
-		}
-
-		p.errors = append(p.errors, fmt.Errorf("%d:%d: expected ';' or newline after struct field, got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-
-		p.consumeTo(token.RBrace)
-	}
-	rbrace := p.expect(token.RBrace, "expected '}'")
-	st.RBrace = rbrace
-
-	return st
-}
-
-// parseStructTypeField is in charge of parsing struct field
-func (p *Parser) parseStructTypeField() *ast.FieldDecl {
-	kw := p.expect(token.Ident, "expected 'ident'")
-	fd := &ast.FieldDecl{
-		Name:   kw,
-		Public: isPublic(kw),
-	}
-	tok := p.peek()
-
-	if !token.IsStructFieldTypes(tok.Kind) {
-		fd.Type = &ast.BadType{From: tok, Reason: "unexpected type name"}
-		p.errors = append(p.errors, fmt.Errorf("%d:%d: unsupported type with %v %q", tok.Line, tok.Column, tok.Kind, tok.Value))
-		p.consumeTo(token.EOF)
-		return fd
-	}
-	fd.Type = &ast.NameType{Name: tok}
-	_ = p.next()
-
-	if p.kind() == token.Assign {
-		kwa := p.expect(token.Assign, "expected '='")
-		fd.Eq = &kwa
-		fd.Default = p.parseExpr(LOWEST)
-	}
-	return fd
-}
-
 // isPublic returns if the field is public or not
 func isPublic(z token.Token) bool {
 	ch := z.Value[0]
 	return ch >= 'A' && ch <= 'Z'
-}
-
-// parseInterfaceType returns parsed interface
-func (p *Parser) parseInterfaceType() *ast.InterfaceType {
-	kwt := p.expect(token.KWType, "expected 'type'")
-	kwi := p.expect(token.Ident, "expected 'ident'")
-	kws := p.expect(token.KWInterface, "expected 'interface'")
-	lbrace := p.expect(token.LBrace, "expected '{'")
-
-	it := &ast.InterfaceType{
-		TypeDecl:  kwt,
-		Name:      kwi,
-		Public:    isPublic(kwi),
-		Interface: kws,
-		LBrace:    lbrace,
-	}
-
-	for p.kind() != token.RBrace && p.kind() != token.EOF {
-		if p.kind() == token.Ident {
-			if p.kindNext(p.position+1) == token.LParen {
-				it.Methods = append(it.Methods, p.parseFuncSignature())
-			} else {
-				it.Embeds = append(it.Embeds, p.parseInterfaceTypeEmbbed())
-			}
-		}
-
-		if p.kind() == token.Comment {
-			_ = p.next()
-		}
-
-		if p.kind() == token.SemiComma {
-			_ = p.next()
-			continue
-		}
-
-		if p.kind() == token.RBrace {
-			break
-		}
-
-		if p.newlineSincePrev() {
-			continue
-		}
-
-		p.errors = append(p.errors, fmt.Errorf("%d:%d: expected ';' or newline after interface field, got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-
-		p.consumeTo(token.RBrace)
-	}
-
-	rbrace := p.expect(token.RBrace, "expected '}'")
-	it.RBrace = rbrace
-
-	return it
-}
-
-// parseFuncSignature returns function signature for interface
-func (p *Parser) parseFuncSignature() ast.InterfaceMethods {
-	name := p.expect(token.Ident, "expected function name")
-	_ = p.expect(token.LParen, "expected '(' after function name")
-
-	f := ast.InterfaceMethods{
-		Name: name,
-	}
-	for p.kind() != token.RParen && p.kind() != token.EOF {
-		if p.kind() == token.Comma {
-			tok := p.next()
-			p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", tok.Line, tok.Column, tok.Kind, tok.Value))
-			p.consumeTo(token.Comma)
-			return f
-		}
-		f.Params = append(f.Params, p.parseFuncSignatureParam())
-		if p.kind() != token.Comma && p.kind() != token.RParen && p.kind() != token.EOF {
-			tok := p.next()
-			p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", tok.Line, tok.Column, tok.Kind, tok.Value))
-			p.consumeTo(token.Comma)
-			return f
-		}
-
-		if p.kind() == token.Comma {
-			_ = p.next()
-			if p.kind() == token.RParen || p.kind() == token.EOF {
-				p.errors = append(p.errors, fmt.Errorf("%d:%d: unexpected expression, got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-				p.consumeTo(token.Comma)
-				return f
-			}
-		}
-	}
-
-	_ = p.expect(token.RParen, "expected ')' after function name")
-	f.Results = p.parseFuncSignatureReturnTypes()
-
-	return f
-}
-
-// parseFuncSignatureParam returns function parameter
-func (p *Parser) parseFuncSignatureParam() ast.Param {
-	name := p.expect(token.Ident, "expected parameter identifier")
-	return ast.Param{Name: name, Type: p.parseFuncSignatureParamType()}
-}
-
-// parseFuncSignatureParamType returns func parameter type
-func (p *Parser) parseFuncSignatureParamType() *ast.NameType {
-	typ := &ast.NameType{}
-	tok := p.next()
-
-	if token.IsFuncParamTypes(tok.Kind) {
-		typ.Name = tok
-	} else {
-		p.errors = append(p.errors, fmt.Errorf("%d:%d: unsupported type with %v %q", tok.Line, tok.Column, tok.Kind, tok.Value))
-		p.consumeTo(token.RParen)
-	}
-	return typ
-}
-
-// parseFuncSignatureReturnTypes returns func return types
-func (p *Parser) parseFuncSignatureReturnTypes() ast.ReturnTypes {
-	var result ast.ReturnTypes
-	if p.kind() == token.LParen {
-		lp := p.expect(token.LParen, "expected '('")
-		if p.kind() == token.RParen {
-			p.errors = append(p.errors, fmt.Errorf("%d:%d: expected parameter(s) before ')', got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-			result.List = append(result.List, ast.Param{Type: p.parseFuncSignatureParamType()})
-			return result
-		}
-
-		if p.kind() == token.Comma {
-			p.errors = append(p.errors, fmt.Errorf("%d:%d: expected expression before ',', got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-
-			result.List = append(result.List, ast.Param{Type: p.parseFuncSignatureParamType()})
-			return result
-		}
-
-		result.LParen = lp
-		// entering into kind: (indentA indentB, indentC indentD) or (indentA indentB)
-		if p.kindNext(p.position+2) == token.Comma || p.kindNext(p.position+2) == token.RParen {
-			for p.kind() != token.RParen && p.kind() != token.LBrace && p.kind() != token.EOF {
-				result.List = append(result.List, p.parseFuncSignatureParam())
-
-				if p.kind() != token.Comma && p.kind() != token.RParen {
-					p.errors = append(p.errors, fmt.Errorf("%d:%d: expected ',' after parameter(s), got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-					return result
-				}
-
-				if p.kind() == token.Comma {
-					_ = p.expect(token.Comma, "expected ','")
-					if p.kind() == token.RParen {
-						p.errors = append(p.errors, fmt.Errorf("%d:%d: expected parameter(s) after ',', got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-						return result
-					}
-				}
-			}
-		} else {
-			// entering into kind: (type, type)
-			for p.kind() != token.RParen && p.kind() != token.LBrace && p.kind() != token.EOF {
-				result.List = append(result.List, ast.Param{Type: p.parseFuncSignatureParamType()})
-
-				if p.kind() != token.Comma && p.kind() != token.RParen {
-					p.errors = append(p.errors, fmt.Errorf("%d:%d: expected ',' after parameter(s), got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-					return result
-				}
-
-				if p.kind() == token.Comma {
-					_ = p.expect(token.Comma, "expected ','")
-					if p.kind() == token.RParen {
-						p.errors = append(p.errors, fmt.Errorf("%d:%d: expected parameter(s) after ',', got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-						return result
-					}
-				}
-			}
-		}
-
-		rp := p.expect(token.RParen, "expected ')'")
-		result.RParen = rp
-		return result
-	}
-
-	if p.kind() == token.Comma {
-		p.errors = append(p.errors, fmt.Errorf("%d:%d: expected builtin type or ident, got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-		return result
-	}
-
-	result.List = append(result.List, ast.Param{Type: p.parseFuncSignatureParamType()})
-	return result
-}
-
-// parseInterfaceTypeEmbbed returns embbed signature for interface
-func (p *Parser) parseInterfaceTypeEmbbed() ast.TypeRef {
-	kw := p.expect(token.Ident, "expected 'ident'")
-
-	tr := ast.TypeRef{}
-	tr.Parts = append(tr.Parts, kw)
-	if p.kind() == token.Dot {
-		tr.Parts = append(tr.Parts, p.next())
-		if p.kind() != token.Ident {
-			p.errors = append(p.errors, fmt.Errorf("%d:%d: expected ident after '.', got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-			p.consumeTo(token.RBrace)
-			return tr
-		}
-		tr.Parts = append(tr.Parts, p.next())
-	}
-	return tr
-}
-
-// parseImplementsDecl is in charge of parsing interface "implements" requirements
-func (p *Parser) parseImplementsDecl() *ast.ImplementsDecl {
-	kw := p.expect(token.Ident, "expected 'ident'")
-	kwi := p.expect(token.KWImplements, "expected 'implements'")
-	id := &ast.ImplementsDecl{
-		Type:       kw,
-		Implements: kwi,
-	}
-
-	if p.kind() != token.Ident {
-		p.errors = append(p.errors, fmt.Errorf("%d:%d: expected ident after 'implements', got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-		p.consumeTo(token.RBrace)
-		return id
-	}
-
-	id.Interface = p.parseInterfaceTypeEmbbed()
-	if p.kind() == token.SemiComma {
-		_ = p.next()
-	}
-	return id
-}
-
-// parseEnumDecl is in charge of parsing enum type
-func (p *Parser) parseEnumDecl() *ast.EnumType {
-	kwp := p.expect(token.KWType, "expected 'type'")
-	kwi := p.expect(token.Ident, "expected 'ident'")
-	kwe := p.expect(token.KWEnum, "expected 'enum'")
-	kweq := p.expect(token.Assign, "expected '='")
-
-	ed := &ast.EnumType{
-		TypeDecl: kwp,
-		Name:     kwi,
-		Public:   isPublic(kwi),
-		Eq:       kweq,
-		Enum:     kwe,
-	}
-
-	if p.kind() != token.Pipe {
-		p.errors = append(p.errors, fmt.Errorf("%d:%d: expected '| after '=', got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-		p.consumeTo(token.EOF)
-		return ed
-	}
-
-	for p.kind() != token.EOF {
-		if p.kind() == token.Pipe {
-			_ = p.next()
-		}
-
-		if p.kind() != token.Ident {
-			p.errors = append(p.errors, fmt.Errorf("%d:%d: expected ident after '|', got %v %q", p.peek().Line, p.peek().Column, p.peek().Kind, p.peek().Value))
-			p.consumeTo(token.EOF)
-		}
-
-		ed.Variants = append(ed.Variants, p.next())
-
-		if p.kind() == token.Comment {
-			_ = p.next()
-		}
-
-		if p.kind() != token.Pipe {
-			break
-		}
-	}
-
-	return ed
 }
